requirements.txt
# Python dependencies
flask>=2.3.0,<3.0.0
folium>=0.14.0
groq>=0.4.0
pandas>=2.0.0
requests>=2.31.0
geopandas>=0.13.0
python-dotenv>=1.0.0
flask-cors>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
flake8>=6.0.0
black>=23.0.0
bandit>=1.7.5

# Security
safety>=2.3.0
